+++ src/core_modules/capture-ui/FlatList/FlatList.component.js
@@ -1,13 +1,50 @@
 // @flow
-import React from 'react';
-import uuid from 'uuid/v4';
+import React, { type ComponentType } from 'react';
+import cx from 'classnames';
+import { colors, spacersNum } from '@dhis2/ui';
+import { withStyles } from '@material-ui/core';
 import type { Props } from './flatList.types';
 
-export const FlatList = ({ list, className, dataTest, renderItem }: Props) => (
-    <div
-        data-test={dataTest}
-        className={className}
-    >
-        {list.map(item => (<div key={uuid()}>{renderItem(item)}</div>))}
-    </div>
-);
+
+const styles = {
+    flatListWrapper: {
+        padding: `0 ${spacersNum.dp16}px`,
+    },
+    itemRow: {
+        borderBottom: `1px solid${colors.grey400}`,
+        display: 'flex',
+        padding: `${spacersNum.dp16}px 0`,
+        '&.isLastItem': {
+            borderBottomWidth: 0,
+        },
+    },
+    itemKey: {
+        width: 150,
+        color: colors.grey600,
+    },
+};
+
+const FlatListPlain = ({ list, classes, dataTest }: Props) => {
+    const lastItemKey = list[list.length - 1]?.reactKey;
+    const renderItem = item => (
+        <div
+            key={item.reactKey}
+            className={cx(classes.itemRow, { isLastItem: item.reactKey === lastItemKey })}
+        >
+            <div className={classes.itemKey}>{item.key}</div>
+            <div>{item.value}</div>
+        </div>
+    );
+
+    return (
+        <div
+            data-test={dataTest}
+            className={classes.flatListWrapper}
+        >
+            {list.map(item => renderItem(item))}
+        </div>
+    );
+};
+
+
+export const FlatList: ComponentType<$Diff<Props, CssClasses>> = withStyles(styles)(FlatListPlain);